requirements.txt
smbus
board
adafruit_dht
psutil
psycopg2
paho-mqtt
